Add URL searching, refactor linking, fix crash
Related #608
