+++ src/features/search/SearchOptions.tsx
@@ -3,10 +3,13 @@ import { InsetIonItem, SettingLabel } from "../user/Profile";
 import { useBuildGeneralBrowseLink } from "../../helpers/routes";
 import {
   albumsOutline,
+  arrowForward,
   chatbubbleOutline,
   personOutline,
   searchOutline,
 } from "ionicons/icons";
+import useLemmyUrlHandler from "../shared/useLemmyUrlHandler";
+import { useMemo } from "react";
 
 interface SearchOptionsProps {
   search: string;
@@ -14,13 +17,30 @@ interface SearchOptionsProps {
 
 export default function SearchOptions({ search }: SearchOptionsProps) {
   const buildGeneralBrowseLink = useBuildGeneralBrowseLink();
+  const { determineObjectTypeFromUrl, redirectToLemmyObjectIfNeeded } =
+    useLemmyUrlHandler();
 
   const searchURI = encodeURIComponent(search);
 
   const sanitizedUser = search.replace(/|\/|#|\?|\\/g, "").replace(/^@/, "");
 
+  const type = useMemo(
+    () => determineObjectTypeFromUrl(search),
+    [determineObjectTypeFromUrl, search],
+  );
+
   return (
     <IonList inset color="primary">
+      {type && (
+        <InsetIonItem
+          onClick={(e) => redirectToLemmyObjectIfNeeded(search, e)}
+          detail
+          button
+        >
+          <IonIcon icon={arrowForward} color="primary" />
+          <SettingLabel>Visit {type}</SettingLabel>
+        </InsetIonItem>
+      )}
       <InsetIonItem routerLink={`/search/posts/${searchURI}`}>
         <IonIcon icon={albumsOutline} color="primary" />
         <SettingLabel>Posts with “{search}”</SettingLabel>
